refactor: Rename WorkoutNavigation.kt
Rename WorkoutNavigation.kt to TrainingHistoryNavigation.kt, since
it is more consistent with the naming of other training-related
navigation files
